requirements.txt
# FastAPI y servidor
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Base de datos
sqlalchemy>=2.0.0
pyodbc>=5.0.0

# IA y ML
faster-whisper>=0.9.0
google-genai>=0.2.0

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.0.0

protobuf